// apps/api/src/tasks/tasks.model.ts
export class AirgradientModel {
  locationId: number;
  locationName: string;
  publicLocationName: string;
  publicPlaceName: string;
  publicPlaceUrl: string;
  publicContributorName: string;
  latitude: number;
  longitude: number;
  offline: boolean;
  pm01: number;
  pm02: number;
  pm10: number;
  pm003Count: number;
  atmp: number;
  rhum: number;
  rco2: number;
  tvoc: number;
  tvocIndex: number;
  noxIndex: number;
  heatindex: number;
  timestamp: string;
  timezone: string;
  model: string;
  firmwareVersion: string;
  wifi: number;
}